introduce a new crate for trace handler
