+++ crates/trace-handler/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "air-trace-handler"
+version = "0.1.0"
+description = "Implementation of AIR trace handler"
+authors = ["Fluence Labs"]
+edition = "2018"
+license = "Apache-2.0"
+repository = "https://github.com/fluencelabs/air/crates/trace-handler"
+publish = false
+keywords = ["fluence", "air", "webassembly", "programming-language"]
+categories = ["wasm"]
+
+[lib]
+name = "air_trace_handler"
+path = "src/lib.rs"
+
+[dependencies]
+air-interpreter-data = { path = "../interpreter-data" }
+air-parser = { path = "../air-parser" }
+
+serde_json = "1.0.68"
+log = "0.4.14"
+thiserror = "1.0.29"
