+++ crates/trace-handler/src/state_automata/fold_fsm/lore_ctor.rs
