+++ crates/trace-handler/src/state_automata/par_fsm/par_builder.rs
