+++ crates/trace-handler/src/state_automata/state_inserter.rs
